requirements.txt
django
psycopg2
djangorestframework
requests
djangorestframework-jwt